Added new languages
