+++ src/i18n/__texts/words.js
@@ -1,0 +1,211 @@
+/**
+ * NOTE: separate text objects if it gets too big in size
+ * bec the translator has a limit (unspecified in docs)
+ */
+const words = [
+  {
+    Hedge: "Hedge",
+    Long: "Long",
+    Crypto: "Crypto",
+    Fiat: "Fiat",
+    Accept: "Accept",
+    Account: "Account",
+    All: "All",
+    Add: "Add",
+    Amount: "Amount",
+    Address: "Address",
+    Applications: "Applications",
+    Approve: "Approve",
+    Apps: "Apps",
+    Balance: "Balance",
+    Biometric: "Biometric",
+    Bridge: "Bridge",
+    Back: "Back",
+    Branch: "Branch",
+    Cancel: "Cancel",
+    City: "City",
+    CashTokens: "CashTokens",
+    Clear: "Clear",
+    Close: "Close",
+    Country: "Country",
+    Collectibles: "Collectibles",
+    Connect: "Connect",
+    Confirmation: "Confirmation",
+    Continue: "Continue",
+    Copying: "Copying",
+    Create: "Create",
+    Currency: "Currency",
+    Date: "Date",
+    Dismiss: "Dismiss",
+    Data: "Data",
+    Decoded: "Decoded",
+    Devices: "Devices",
+    Deadline: "Deadline",
+    Deposit: "Deposit",
+    Disconnected: "Disconnected",
+    English: "English",
+    Error: "Error",
+    Done: "Done",
+    From: "From",
+    Gas: "Gas",
+    German: "German",
+    Home: "Home",
+    Keep: "Keep",
+    Language: "Language",
+    Link: "Link",
+    Location: "Location",
+    Manage: "Manage",
+    Memo: "Memo",
+    Message: "Message",
+    MAX: "MAX",
+    Name: "Name",
+    or: "or",
+    Pin: "PIN",
+    Personalize: "Personalize",
+    ChangePin: "Change PIN",
+    POSID: "POSID",
+    Ramp: "Ramp",
+    Receive: "Receive",
+    Received: "Received",
+    Recipient: "Recipient",
+    Recipients: "Recipients",
+    Refetch: "Refetch",
+    Reject: "Reject",
+    Remove: "Remove",
+    Rename: "Rename",
+    Required: "Required",
+    Raw: "Raw",
+    Retry: "Retry",
+    Route: "Route",
+    Routing: "Routing",
+    Scan: "Scan",
+    Scanning: "Scanning",
+    Security: "Security",
+    Send: "Send",
+    Sender: "Sender",
+    Senders: "Senders",
+    Sent: "Sent",
+    Settings: "Settings",
+    Setup: "Setup",
+    Set: "Set",
+    Skip: "Skip",
+    Slippage: "Slippage",
+    Spanish: "Spanish",
+    Street: "Street",
+    Success: "Success",
+    Swap: "Swap",
+    Sweep: "Sweep",
+    Symbol: "Symbol",
+    To: "To",
+    Tokens: "Tokens",
+    Transactions: "Transactions",
+    Update: "Update",
+    Verify: "Verify",
+    Value: "Value",
+    Version: "Version",
+    Wallet: "Wallet",
+    Wallets: "Wallets",
+    Details: "Details",
+    Transaction: "Transaction",
+    Extensions: "Extensions",
+    Description: "Description",
+    Unlink: "Unlink",
+    Gifts: "Gifts",
+    Commitment: "Commitment",
+    Capability: "Capability",
+    Submit: "Submit",
+    Select: "Select",
+    Duration: "Duration",
+    Low: "Low",
+    Asset: "Asset",
+    Calculate: "Calculate",
+    Expired: "Expired",
+    High: "High",
+    Liquidity: "Liquidity",
+    ArgentinianSpanish: "Spanish (Argentina)",
+    BrazilianPortuguese: "Portuguese (Brazil)",
+    Hausa: "Hausa",
+  },
+  {
+      Type: "Type",
+      Refund: "Refund",
+      Arbitrary: "Arbitrary",
+      Total: "Total",
+      Submitted: "Submitted",
+      Resubmit: "Resubmit",
+      Liquidation: "Liquidation",
+      Settlement: "Settlement",
+      Start: "Start",
+      Summary: "Summary",
+      Maturation: "Maturation",
+      Loss: "loss",
+      Gain: "gain",
+      Signed: "Signed",
+      Pending: "Pending",
+      Decline: "Decline",
+      Minute: "minute",
+      Hour: "hour",
+      Day: "day",
+      Week: "week",
+      TildeMonth: "~month",
+      TildeMonths: "~months",
+      TildeYear: "~year",
+      TildeYears: "~years",
+      By: 'by',
+      Accepted: "Accepted",
+      Agreed: "Agreed",
+      Removing: "Removing",
+      Removed: "Removed",
+      Claim: "Claim",
+      Generate: "Generate",
+      Proceed: "Proceed",
+      Recover: "Recover",
+      Claimed: "Claimed",
+      Unclaimed: "Unclaimed",
+      Campaign: "Campaign",
+      Recovered: "Recovered",
+      Ago: "ago",
+      Seconds: "seconds",
+      Second: "second",
+      Allowance: "Allowance",
+      Range: "Range",
+      Monthly: "Monthly",
+      Daily: "Daily",
+      Filter: "Filter",
+      OK: "OK",
+      Properties: "Properties",
+      BranchSmall: "branch",
+      Branches: "branches",
+      Stabilize: "Stabilize",
+      Leverage: "Leverage",
+      Funding: "Funding",
+      Contract: "Contract",
+      Fees: "Fees",
+      Premium: "Premium",
+      Units: "units",
+      Session: "Session",
+      Method: "Method",
+      Chain: "Chain",
+      Requests: "Requests",
+      Topic: "Topic",
+      Connecting: "Connecting",
+      Minutes: "minutes",
+      Hours: "hours",
+      Days: "days",
+      Weeks: "weeks",
+      Months: "months",
+      Confirm: "Confirm",
+      Offer: "offer",
+      Warning: "Warning",
+      Yes: "YES",
+      No: "NO",
+      To: "to",
+      Recipient: "Recipient",
+      Points: "Points",
+      DEEM: "DEEM",
+      Theme: "Theme",
+      Default: "Default",
+  },
+]
+
+module.exports = words
