+++ src/i18n/af/index.js
@@ -3,11 +3,11 @@
 
 export default {
   AboutTheApp: "Sobre o aplicativo",
-  Accept: "Aceitar",
+  Accept: "Aanvaar",
   Accepted: "Aceitaram",
-  Account: "Conta",
+  Account: "Rekening",
   ActiveSessions: "Sessões ativas",
-  Add: "Adicionar",
+  Add: "Byvoeg",
   AddAll: "Adicione tudo",
   AddAnotherRecipient: "Adicione outro destinatário",
   AddFungibleCashPoint: "Adicione o caixa fungível",
@@ -20,19 +20,19 @@ export default {
   Add_Type1_Point: "Adicionar SLP tipo 1 ponto",
   Add_Type1_Token: "Adicionar slp tipo 1 token",
   AddingNewDevice: "Adicionando novo dispositivo",
-  Address: "Endereço",
+  Address: "Toespraak",
   AddressScan: "Digitalização de endereço",
   Ago: "atrás",
   Agreed: "Acordado",
-  All: "Todos",
+  All: "Almal",
   Allowance: "Mesada",
-  Amount: "Quantia",
+  Amount: "Bedrag",
   AmountGreaterThanBalance: "O valor é maior que o seu equilíbrio",
   AnyHedgeContract: "AnyendeDGECONTRATRA",
   AnyHedgeNoExactMatchInfo: "Se não houver uma correspondência exata encontrada no pool, uma lista de ofertas semelhantes é sugerida",
   AppInfo: "Informações do aplicativo",
-  Applications: "Formulários",
-  Approve: "Aprovar",
+  Applications: "Aansoeke",
+  Approve: "Goedkeur",
   ApprovePoint: "Aprovar ponto",
   ApproveToken: "Aprovar o token",
   ApprovingPoint: "Ponto de aprovação",
@@ -41,21 +41,21 @@ export default {
   ApprovingTokenName: "Aprovação {tokenInfoName}",
   ApproxHedgeAmount: "Quantidade de hedge de aprox",
   ApproxHedgeAmountDescription: "A quantidade de hedge é calculada a partir de uma longa quantidade abaixo e a baixa porcentagem de liquidação abaixo",
-  Apps: "Aplicativos",
+  Apps: "Programme",
   Arbitrary: "Arbitrário",
   AreYouSure: "Tem certeza?",
-  ArgentinianSpanish: "Espanhol (Argentina)",
-  Asset: "Ativo",
+  ArgentinianSpanish: "Spaans (Argentinië)",
+  Asset: "Aanvaarding",
   AssetName: "Nome do ativo",
   AssetRemovalText: "Quer continuar removendo o ativo?",
   AssetSwap: "Troca de ativos",
   AssetValueNote: "O valor do ativo é baseado nos preços no momento da transação",
   AssetsSwappedSuccesfully: "Ativos trocados com sucesso",
   BCHBullDescription: "Crie contratos de Anyhedge instantaneamente com um provedor de liquidez",
-  Back: "Voltar",
+  Back: "Agter",
   BackToEdit: "Voltar á editar",
   BackendDown: "Nosso servidor de back -end é inacessível.Isso pode ser devido à sua conexão com a Internet ou nosso servidor sendo temporariamente baixo",
-  Balance: "Equilíbrio",
+  Balance: "Balans",
   BalanceExceeded: "Saldo excedido",
   BchAddress: "Endereço BCH",
   BchAddresses: "Endereços BCH",
@@ -63,17 +63,17 @@ export default {
   BchBridge: "Ponte BCH",
   BchToReceive: "BCH para receber",
   BchToSend: "BCH para enviar",
-  Biometric: "Biométrico",
+  Biometric: "Biometries",
   BiometricMaxAttemptsMsg: "Você \"fez muitas tentativas. Por favor, tente novamente após 30 segundos.",
-  Branch: "Filial",
+  Branch: "Tak",
   BranchDetails: "Detalhes da filial",
   BranchDetailsSaved: "Detalhes da filial salvos",
   BranchName: "Nome da filial",
   BranchRemoved: "Ramo removido",
   BranchSmall: "filial",
   Branches: "galhos",
-  BrazilianPortuguese: "Português (Brasil)",
-  Bridge: "Ponte",
+  BrazilianPortuguese: "Portugees (Brasilië)",
+  Bridge: "Brug",
   BridgeBalance: "Equilíbrio da ponte",
   BridgeDisabled: "A ponte está temporariamente desativada até novo aviso",
   BridgeError1: "Deve ser pelo menos 0,01",
@@ -83,7 +83,7 @@ export default {
   BusinessName: "Nome da empresa",
   By: "por",
   ByA: "por a",
-  Calculate: "Calcular",
+  Calculate: "Bereken",
   CalculateFeesError: "Erro calculando taxas de contrato",
   CalculatingContractFees: "Cálculo de taxas de contrato",
   CalculatingFundingAmount: "Calcular o valor do financiamento",
@@ -99,7 +99,7 @@ export default {
   CampaignLimitError: "O limite de campanha por carteira não pode ser maior que o valor do presente",
   CampaignName: "Nome da campanha",
   CampaignOptional: "Campanha (opcional)",
-  Cancel: "Cancelar",
+  Cancel: "Kanselleer",
   CancelContract: "Cancelar contrato",
   CancelMutualRedemptionProposal: "Cancelar proposta de resgate mútua",
   CancelProposal: "Cancelar proposta",
@@ -110,29 +110,29 @@ export default {
   CancellingUnlinkRequest: "Cancelar a solicitação desvinculada",
   CannotAddRecipient: "Não pode adicionar mais de 5 destinatários.",
   CannotBeLowerThanGiftAmount: "Isso não pode ser menor que o valor do presente",
-  Capability: "Capacidade",
-  CashTokens: "CashTokens",
+  Capability: "Vermoë",
+  CashTokens: "Cashtokens",
   CategoryID: "Categoria ID",
   Chain: "Corrente",
-  ChangePin: "Alterar pino",
+  ChangePin: "Verander speld",
   ChineseSimplified: "Chinês simplificado",
   ChineseTraditional: "Chinês tradicional",
   ChoosePreferredSecAuth: "Escolha sua autenticação de segurança preferida",
-  City: "Cidade",
+  City: "Stad",
   Claim: "Alegar",
   ClaimGift: "Presente de reivindicação",
   ClaimGiftCompleted: "Presente de reclamação concluído!",
   Claimed: "Reivindicado",
   ClaimingGift: "Reivindicação do presente ...",
-  Clear: "Claro",
+  Clear: "Helder",
   ClearCallRequests: "Solicitações de chamada clara",
   ClearCallRequestsPrompt: "Removendo todas as solicitações de chamada.Tem certeza?",
   ClearFilter: "Filtro limpo",
   ClickToCopyAddress: "Clique para copiar o endereço",
   ClickToReveal: "Clique para revelar",
-  Close: "Fechar",
-  Collectibles: "Colecionáveis",
-  Commitment: "Compromisso",
+  Close: "Naby",
+  Collectibles: "Versamelstukke",
+  Commitment: "Toewyding",
   CompilingContractError: "Erro encontrado na compilação de contrato",
   CompleteFundingProposal: "Proposta de financiamento completo",
   CompletingContractFunding: "Concluindo o financiamento do contrato",
@@ -141,10 +141,10 @@ export default {
   ConfirmMutualRedemptionProposal: "Confirmar proposta de resgate mútua",
   ConfirmPin: "Confirme o pino",
   ConfirmSwap: "Confirme a troca",
-  Confirmation: "Confirmação",
-  Connect: "Conectar",
+  Confirmation: "Bevestiging",
+  Connect: "Konnekteer",
   Connecting: "Conectando",
-  Continue: "Continuar",
+  Continue: "Aanhou",
   Contract: "Contrato",
   ContractAddressMismatch: "Endereço do contrato incompatibilidade.Incapaz de criar financiamento UTXO",
   ContractCancelled: "Contrato cancelado",
@@ -159,9 +159,9 @@ export default {
   ContractWasCancelled: "Contrato foi cancelado",
   CopiedToClipboard: "Copiado para a área de transferência",
   CopyPrivateKey: "Copie a chave privada",
-  Copying: "Cópia",
-  Country: "País",
-  Create: "Criar",
+  Copying: "Kopieer",
+  Country: "Nasie",
+  Create: "Skep",
   CreateContractError: "Erro encontrado criando contrato",
   CreateGift: "Crie presente",
   CreateHedge: "Criando oferta de posição de hedge",
@@ -177,33 +177,33 @@ export default {
   CreatingUTXO: "Criando UTXO para proposta de financiamento",
   CreatingUnlinkDeviceRequest: "Criando solicitação de dispositivo desvinculada",
   CreatingYourWallet: "Criando sua carteira",
-  Crypto: "Cripto",
-  Currency: "Moeda",
+  Crypto: "Kripto",
+  Currency: "Geldstelsel",
   CurrentPrice: "Preço atual",
   CurrentTimeError: "Deve ser após o horário atual",
   DEEM: "Deem",
   Daily: "Diário",
   DarkMode: "Modo escuro",
-  Data: "Dados",
-  Date: "Data",
+  Data: "Data",
+  Date: "Datum",
   DateCreated: "Data Criada",
   DateFrom: "Data de",
   DateTo: "Data para",
   Day: "dia",
   Days: "dias",
-  Deadline: "Prazo final",
+  Deadline: "Sperdatum",
   Decline: "Declínio",
   DeclineMutualRedemptionProposal: "Proposta de resgate mútua em declínio",
   DeclineProposal: "Declinar proposta",
-  Decoded: "Decodificado",
+  Decoded: "Gedekodeer",
   Default: "Padrão",
   DeleteWallet: "Excluir carteira",
   DeleteWalletDescription: "Tem certeza de que deseja excluir esta carteira?",
   DeleteWalletNow: "Exclua a carteira agora",
-  Deposit: "Depósito",
+  Deposit: "Deposito",
   DerivationPath: "Caminho de derivação",
-  Description: "Descrição",
-  Details: "Detalhes",
+  Description: "Beskrywing",
+  Details: "Besonderheid",
   DetectedUnknownCurrency: "Detecida moeda desconhecida: #{currency}",
   DetermineProposalError: "Incapaz de determinar a posição para verificar",
   DeviceAddedIDNo: "Dispositivo adicionado #{ID}",
@@ -212,15 +212,15 @@ export default {
   DeviceSuspended: "Dispositivo suspenso",
   DeviceSuspendedPOS: "O dispositivo está atualmente suspenso",
   DeviceUnsuspended: "Dispositivo não suspenso",
-  Devices: "Dispositivos",
-  Disconnected: "Desconectado",
-  Dismiss: "Liberar",
-  Done: "Feito",
-  Duration: "Duração",
+  Devices: "Toestelle",
+  Disconnected: "Ontkoppel",
+  Dismiss: "Afdank",
+  Done: "Gedoen",
+  Duration: "Duur",
   EarlyMaturation: "Maturação precoce",
   EmptyRecipient: "Destinatário não pode estar vazio",
   EnableSmartBCH: "Ativar SmartBCH",
-  English: "Inglês",
+  English: "Engels",
   EnterAmount: "Insira o valor",
   EnterCashPointCategoryID: "Digite ID da categoria de CashPoint",
   EnterCashTokenCategoryID: "Digite ID da categoria CashToken",
@@ -230,14 +230,14 @@ export default {
   Enter_SLP_PointId: "Digite SLP Point ID",
   Enter_SLP_TokenId: "Digite ID do token SLP",
   ErrAcceptingCallRequest: "Erro aceitar a solicitação de chamada",
-  Error: "Erro",
+  Error: "Fout",
   EstimatedFee: "Taxa estimada",
   ExceededBalanceError: "Exceder o equilíbrio {spendableBchValue}",
   ExpirationUpdated: "Expiração atualizada",
-  Expired: "Expirado",
+  Expired: "Verstryk",
   ExpiresAt: "Expira em",
   ExplorerLink: "Link Explorer",
-  Extensions: "Extensões",
+  Extensions: "Uitbreidings",
   FailedAddingNewDevice: "Falha ao adicionar novo dispositivo",
   FailedRemoveBranch: "Falha ao remover o ramo",
   FailedToCreateRandomSig: "Falhou em criar assinatura aleatória",
@@ -259,7 +259,7 @@ export default {
   FindingMatchingPositionError: "Erro ao encontrar a correspondência {hedgeLongPosition} posição",
   FindingUnlistedAssets: "Encontrando ativos não listados",
   FoundExistingOffer: "Encontrou oferta existente.Aceitar oferta de posição",
-  From: "De",
+  From: "Van",
   FundHedgePosition: "Posição de hedge de fundos",
   FundLongPosition: "Financiar uma posição longa",
   Funding: "Financiamento",
@@ -271,7 +271,7 @@ export default {
   FundingTransactionSubmitted: "Transação de financiamento enviada!",
   FundingValidationFailed: "A verificação da proposta de financiamento falhou",
   Gain: "ganho",
-  Gas: "Gás",
+  Gas: "Gas",
   GasFee: "Taxa de gás",
   GasPrice: "Preço do gás",
   Generate: "Gerar",
@@ -281,15 +281,15 @@ export default {
   GeneratingAddressSet: "Gerando conjunto de endereços",
   GeneratingAddressesError: "Erro de geração de endereços",
   GeneratingRandomSignature: "Gerando assinatura aleatória",
-  German: "Alemão",
+  German: "Duitsers",
   GiftAlreadyClaimed: "Este presente foi reivindicado!Tente outro.",
-  Gifts: "Presentes",
+  Gifts: "Geskenke",
   GiftsYouCreated: "Presentes que você criou",
   GoBack: "Volte",
   GreaterThanHundredError: "Deve ser maior que 100%",
   HasBalance: "Tem equilíbrio",
   Hausa: "Hausa",
-  Hedge: "Cobertura",
+  Hedge: "Heining",
   HedgeContract: "Contrato de hedge",
   HedgeOffers: "Ofertas de hedge",
   HedgePayout: "Pagamento de hedge",
@@ -301,9 +301,9 @@ export default {
   HedgePositions: "Posições de hedge",
   HidePoints: "Ocultar pontos",
   HideTokens: "Ocultar fichas",
-  High: "Alto",
+  High: "Hoog",
   HigherThanThousandError: "Não deve ser superior a 1000%",
-  Home: "Lar",
+  Home: "Tuiste",
   Hour: "hora",
   Hours: "horas",
   IgnoredPoints: "Pontos ignorados",
@@ -326,28 +326,28 @@ export default {
   InvalidPhoneNumber: "Número de telefone inválido",
   InvalidRecipient: "O destinatário deve ser um endereço válido #{walletType}",
   IsNewBranch: "É ramo principal",
-  Keep: "Manter",
-  Language: "Linguagem",
+  Keep: "Hou",
+  Language: "Taal",
   LeavingPage: "Página de saída",
   LegacyAddress: "Endereço Legado",
   LegacyAddressWarning: "Você colou um endereço herdado.Certifique-se de que seja um endereço de depósito <span Class = \"highlighted-word\"> BCH </span>, não um endereço de depósito BTC.",
   Leverage: "Aproveitar",
   LightMode: "Modo de luz",
-  Link: "Link",
+  Link: "Skakel",
   LinkExpired: "Link expirado",
   LinkExpiresIn: "O link expira",
   Liquidation: "Liquidação",
   LiquidationParameters: "Parâmetros de liquidação",
   LiquidationPrice: "Preço de liquidação",
-  Liquidity: "Liquidez",
+  Liquidity: "Likiditeit",
   LiquidityLeastError: "A liquidez requer pelo menos {convertedMinimumAmount}",
   LiquidityMostError: "A liquidez requer na maioria {convertedMaximumAmount}",
   LiquidityPool: "Pool de liquidez",
   LoadingMetadata: "Metadados de carregamento",
   LoadingWallet: "Carteira de carregamento ...",
-  Location: "Localização",
+  Location: "Lokasie",
   LogIn: "Conecte-se",
-  Long: "Longo",
+  Long: "Lank",
   LongAmountDescription: "O valor longo é apenas uma aproximação sem um valor de preço inicial no ativo",
   LongOffers: "Longas ofertas",
   LongPayout: "Longo pagamento",
@@ -358,10 +358,10 @@ export default {
   LongPositionOfferRemoved: "Oferta de posição longa removida",
   LongPositions: "Posições longas",
   Loss: "perda",
-  Low: "Baixo",
-  MAX: "Máx",
+  Low: "Laag",
+  MAX: "Maksimum",
   MainBranch: "Ramo principal",
-  Manage: "Gerenciar",
+  Manage: "Bestuur",
   ManageIgnoredPoints: "Gerenciar pontos ignorados",
   ManageIgnoredTokens: "Gerenciar tokens ignorados",
   ManagePoints: "Gerenciar pontos",
@@ -371,11 +371,11 @@ export default {
   Maturation: "Maturação",
   MaxAmountPerWallet: "Quantidade máxima por carteira",
   MaxDuration: "Duração máxima",
-  Memo: "Memorando",
+  Memo: "Memo",
   MerchantDetails: "Detalhes do comerciante",
   MerchantDetailsRequired: "Detalhes do comerciante necessário",
   MerchantDetailsSaved: "Detalhes do comerciante salvos",
-  Message: "Mensagem",
+  Message: "Boodskap",
   Method: "Método",
   MinDuration: "Duração mínima",
   MinerFee: "Taxa de mineiro",
@@ -400,7 +400,7 @@ export default {
   MutualRedemptionProposal: "Proposta de resgate mútua",
   MutualRedemptionSigningError: "Erro de assinatura de resgate mútua",
   MutualRedemptionSubmitted: "Redenção mútua enviada",
-  Name: "Nome",
+  Name: "Naam",
   NativeBiometricReason1: "Para fácil login",
   NativeBiometricReason2: "Para verificação de propriedade",
   NativeBiometricSubtitle: "Verifique sua conta usando impressão digital",
@@ -455,7 +455,7 @@ export default {
   POSDeviceIsNotLinked: "O dispositivo POS não está vinculado",
   POSDeviceMustBeUnlinked: "O dispositivo POS deve ser desvinculado",
   POSDeviceNotFound: "Dispositivo POS não encontrado",
-  POSID: "POSID",
+  POSID: "Posid",
   PasteAddressHere: "Endereço de pasta aqui",
   PasteAddressOrLnsHere: "Colar endereço ou nome LNS aqui",
   PasteGiftCodeHere: "Código de presente colar aqui",
@@ -473,8 +473,8 @@ export default {
   PaytacaMotto: "O TEU DINHEIRO.Seu controle.",
   Pending: "Pendente",
   PendingSessionRequests: "Pedidos de sessão pendente",
-  Personalize: "Personalizar",
-  Pin: "ALFINETE",
+  Personalize: "Personaliseer",
+  Pin: "SPELD",
   PinLocation: "Localização do pino",
   PinMismatched: "PIN incompatível",
   PinSubtext1: "Digite seu PIN para prosseguir.",
@@ -497,18 +497,18 @@ export default {
   Properties: "Propriedades",
   ProposeAnotherRedemption: "Propor outra redenção",
   ProposeMutualRedemption: "Propor resgate mútuo",
-  Ramp: "Rampa",
+  Ramp: "Oprit",
   RampFiatNotice: "Nossa troca entre BCH-to-Fiat, ponto a ponto, estará aqui em breve.Fique atento!",
   Range: "Faixa",
-  Raw: "Cru",
-  Receive: "Receber",
-  Received: "Recebido",
+  Raw: "Rou",
+  Receive: "Ontvang",
+  Received: "Ontvang",
   ReceiverWarningText1: "Primeira vez que recebem fundos para este ativo?Por favor",
   ReceiverWarningText2: "Backup sua frase de semente",
   ReceiverWarningText3: "antes de proceder.",
   ReceivingAddress: "Endereço de recebimento",
-  Recipient: "Destinatário",
-  Recipients: "Destinatários",
+  Recipient: "Ontvanger",
+  Recipients: "Ontvangers",
   Recover: "Recuperar",
   RecoverGift: "Recuperar presente",
   RecoverGiftCompleted: "Recupere o presente concluído!",
@@ -516,15 +516,15 @@ export default {
   Recovered: "Recuperado",
   RecoveringGift: "Presente de recuperação ...",
   ReferenceId: "referência ID",
-  Refetch: "REFETCH",
+  Refetch: "Hervet",
   RefreshList: "Atualizar lista",
   Refund: "Reembolso",
   RefundAddress: "Endereço de reembolso",
   RegisterLongError: "Erro encontrado no registro de contas longas",
-  Reject: "Rejeitar",
+  Reject: "Verwerp",
   RejectedByUser: "Rejeitado pelo usuário",
   RejectedCallRequest: "Solicitação de chamada rejeitada",
-  Remove: "Remover",
+  Remove: "Verwyder",
   RemoveAsset: "Remover ativo",
   RemoveAssetPrompt: "Remova o ativo {assetName}.Tem certeza?",
   RemoveBranch: "Remova o ramo",
@@ -551,10 +551,10 @@ export default {
   RemovingExpiration: "Remoção de expiração",
   RemovingHedgePositionOffer: "Removendo a oferta de posição de hedge",
   RemovingLongPositionOffer: "Removendo oferta de posição longa",
-  Rename: "Renomear",
+  Rename: "Hernoem",
   RenameWallet: "Renomeie a carteira",
   Requests: "solicitações de",
-  Required: "Obrigatório",
+  Required: "Vereis",
   ResolvingLnsAddress: "Resolvendo o endereço de nome do LNS",
   RestoreFromSeedPhrase: "Restaurar da frase de semente",
   RestoreWallet: "Restaurar a carteira",
@@ -566,9 +566,9 @@ export default {
   RetrievePrivateKeyError: "Falha ao recuperar a chave privada",
   RetrievingAddresses: "Recuperação de endereços",
   RetrievingAddressesError: "Erro encontrado na recuperação de endereços",
-  Retry: "Tente novamente",
-  Route: "Rota",
-  Routing: "Roteamento",
+  Retry: "Herver hierdie",
+  Route: "Roete",
+  Routing: "Roete",
   SEP20_VaultBalance: "Balanço do cofre de setembro",
   SLPCollectibles: "SLP Collectibles",
   SLPTokens: "SLP",
@@ -576,26 +576,26 @@ export default {
   SalesReport: "Relatório de vendas",
   SavingLongAccount: "Economizando uma conta longa",
   SavingYourPin: "Salvando seu alfinete",
-  Scan: "Varredura",
+  Scan: "Skandeer",
   ScanClaimGift: "Escanear para reivindicar o presente",
   ScanForPoints: "Digitalize para pontos",
   ScanForTokens: "Digitalize tokens",
   ScanNewSession: "Digitalize a nova sessão",
   ScanQrCode: "Digitalize o código QR",
-  Scanning: "Varredura",
+  Scanning: "Skandering",
   ScanningForUntrackedAddr: "Digitalização para endereços não rastreados",
   ScanningForUtxos: "Digitalização para UTXOS",
   ScanningForUtxosAndAddr: "Digitalização para UTXOS e endereços",
   SearchingForOtherAssets: "Procurando por outros ativos",
   Second: "segundo",
   Seconds: "segundos",
-  Security: "Segurança",
+  Security: "Sekuriteit",
   SecurityAuthentication: "Autenticação de segurança",
   SecurityAuthenticationSetup: "Configuração de autenticação de segurança",
   SecurityCheck: "Verificação de segurança",
   SecurityCheckFailed: "A verificação de segurança falhou",
   SeeBasicWalletInfo: "Veja informações básicas da carteira",
-  Select: "Selecione",
+  Select: "Uitkies",
   SelectActiveSession: "Selecione sessão ativa",
   SelectAssetToBeReceived: "Selecione ativo a ser recebido",
   SelectAssetToSend: "Selecione ativo para enviar",
@@ -612,20 +612,20 @@ export default {
   SelectSource: "Selecione a fonte",
   SelectToken: "Selecione token",
   SelectingFromOffers: "Selecionando de ofertas semelhantes",
-  Send: "Enviar",
+  Send: "Stuur",
   SendAmountGreaterThanZero: "O valor do envio deve ser maior que zero",
   SendPageOffline: "Você não pode enviar fundos enquanto estiver offline.Por favor, conecte -se à Internet.",
   SendToAnotherAddress: "Enviar para outro endereço",
   SendTransaction: "Enviar transação",
-  Sender: "Remetente",
-  Senders: "Remetentes",
-  Sent: "Enviado",
+  Sender: "Sender",
+  Senders: "Senders",
+  Sent: "Gestuur",
   ServiceFee: "Taxa de serviço",
   ServiceFees: "Taxas de serviço",
   Session: "Sessão",
   SessionProposals: "Propostas de sessão",
   SessionURL: "URL da sessão",
-  Set: "Definir",
+  Set: "Dekor",
   SetAmountIn: "Defina o valor em",
   SetCoordinates: "Defina coordenadas",
   SetExpiration: "Defina a expiração",
@@ -633,14 +633,14 @@ export default {
   SetName: "Nome do conjunto",
   SetNewNameForDevice: "Defina um novo nome para o dispositivo",
   SetReceiveAmount: "Definir valor de recebimento",
-  Settings: "Configurações",
+  Settings: "Instellings",
   Settlement: "Povoado",
   SettlementPrice: "Preço de liquidação",
   SettlementService: "Serviço de liquidação",
   SettlementServiceDescription: "Taxa de serviço de liquidação para Paytaca",
   SettlementTransactionNotFound: "Transação de liquidação não encontrada",
   SettlementType: "Tipo de liquidação",
-  Setup: "Configurar",
+  Setup: "Stel op",
   SetupMerchantDetails: "Configurar detalhes do comerciante",
   SetupNewPin: "Configure um novo pino",
   SetupPin: "Pino de configuração",
@@ -663,10 +663,10 @@ export default {
   SigningRequest: "Solicitação de assinatura",
   SimilarOffersError: "Nenhuma oferta correspondente encontrada, mas encontrada ofertas semelhantes.Selecionar um em vez disso?",
   SimilarOffersFound: "Ofertas semelhantes encontradas",
-  Skip: "Pular",
+  Skip: "Huppel",
   SkipVerification: "Pule a verificação",
   SkipVerificationMessage: "Tem certeza de que deseja pular a etapa de verificação?",
-  Slippage: "Deslizamento",
+  Slippage: "Glip",
   SlippageTolerance: "Tolerância a derrapagem",
   SlippageToleranceDescription: "A troca será revertida se o preço mudar desfavoravelmente por essa porcentagem",
   SlpAddress: "Endereço SLP",
@@ -676,25 +676,25 @@ export default {
   SmartSwapBchSoonTokens: "Vamos integrar um DEX em breve!",
   SmartSwapFormErr: "Ocorreu um erro ao buscar informações de troca",
   SourceCodeRepo: "Repositório de código -fonte",
-  Spanish: "Espanhol",
+  Spanish: "Spaans",
   SpendingTransaction: "Transação de gastos",
   Stabilize: "Estabilizar",
   Start: "Começar",
   StartPrice: "Preço de início",
-  Street: "Rua",
-  Submit: "Enviar",
+  Street: "Straat",
+  Submit: "Indien",
   SubmitFundingProposal: "Enviar proposta de financiamento",
   Submitted: "Submetido",
   SubmittingFundingProposal: "Enviando proposta de financiamento",
   SubmittingFundingProposalError: "Erro ao enviar proposta de financiamento",
   SubmittingMutualRedemption: "Enviando resgate mútuo",
   SubmittingMutualRedemptionError: "Erro encontrado ao enviar redenção mútua",
-  Success: "Sucesso",
+  Success: "Sukses",
   SuccessfullySent: "Enviada com sucesso",
   Summary: "Resumo",
   SuspendDevice: "Suspender o dispositivo",
   SuspendingDevice: "Dispositivo de suspensão",
-  Swap: "Trocar",
+  Swap: "Ruil",
   SwapAgain: "Troque novamente",
   SwapFrom: "Trocar de",
   SwapRatio: "Relação de troca",
@@ -702,7 +702,7 @@ export default {
   SwapTo: "Troque para",
   SwapTransactionDeadlineDescription: "A troca será revertida se a transação estiver pendente por mais do que essa duração",
   SwapUpdate: "Atualização de troca",
-  Sweep: "Varrer",
+  Sweep: "Vee",
   SweepErrMsg1: "Este endereço está vazio",
   SweepErrMsg2: "Você precisará de equilíbrio de BCH suficiente para poder varrer o (s) ponto (s) abaixo",
   SweepInputPlaceholder: "Cole aqui a chave privada no formato WIF",
@@ -711,7 +711,7 @@ export default {
   SwipeToConfirm: "Deslizar para confirmar",
   SwipeToSend: "Deslizar para enviar",
   SwitchWallet: "CALLETA DE CURSÃO",
-  Symbol: "Símbolo",
+  Symbol: "Simbool",
   Theme: "Tema",
   ThemeName: "{theme} tema",
   ThemePreferenceSubtitle: "Ajuste o tema para se adequar à sua experiência nativa",
@@ -720,12 +720,12 @@ export default {
   TildeMonths: "~months",
   TildeYear: "~year",
   TildeYears: "~years",
-  To: "para",
+  To: "Na",
   TokenAdded: "Token adicionado",
   TokenAlreadyInList: "Token já existe na lista",
   TokenApproved: "Token aprovado",
   TokenId: "Id token",
-  Tokens: "Tokens",
+  Tokens: "Tekens",
   Topic: "Tema",
   Total: "Total",
   TotalAmountError: "O valor total que está sendo enviado é maior que o seu saldo atual",
@@ -733,12 +733,12 @@ export default {
   TotalLongPositions: "Total de posições longas",
   TotalPayout: "Pagamento total",
   TotalSales: "Vendas totais",
-  Transaction: "Transação",
+  Transaction: "Transaksie",
   TransactionDeadline: "Prazo para transação",
   TransactionHistory: "Histórico de transações",
   TransactionId: "ID da transação",
   TransactionSent: "Transação enviada",
-  Transactions: "Transações",
+  Transactions: "Transaksies",
   Type: "Tipo",
   UTXOScanComplete: "Scan UTXO concluído em",
   UTXOScanOngoing: "UTXO Scan em andamento, começou",
@@ -753,7 +753,7 @@ export default {
   UnknownAsset: "Ativo desconhecido",
   UnknownError: "Erro desconhecido",
   UnknownErrorOccurred: "Ocorreu um erro desconhecido",
-  Unlink: "Desvincular",
+  Unlink: "Skakel",
   UnlinkDevice: "Dispositivo desvinculado",
   UnlinkDeviceRequestCreated: "Solicitação de dispositivo desvinculada criada",
   UnlinkDeviceRequestFailed: "A solicitação de dispositivo desvinculada falhou",
@@ -765,7 +765,7 @@ export default {
   UnresolvedTransactionProposal: "Proposta de transação não resolvida",
   UnsuspendDevice: "Dessúpensa do dispositivo",
   UnsuspendingDevice: "Dispositivo não suspense",
-  Update: "Atualizar",
+  Update: "Opdateer",
   UpdateDeviceIDNo: "Atualizar dispositivo #{ID}",
   UpdateExpiry: "Atualizar expiração",
   UpdatedDeviceIDNo: "Dispositivo atualizado #{ID}",
@@ -779,8 +779,8 @@ export default {
   UtxoScan: "UTXO Scan",
   ValidFundingProposal: "Proposta de financiamento válida",
   ValidatingContractFundingError: "Erro encontrado na validação de financiamento do contrato",
-  Value: "Valor",
-  Verify: "Verificar",
+  Value: "Waarde",
+  Verify: "Bevestig",
   VerifyFundingProposalError: "Erro encontrado na verificação da proposta de financiamento",
   VerifyMnemonicBackupPhrase: "Verifique se Mnemonic embaralhou",
   VerifyPin: "Verifique o pino",
@@ -788,7 +788,7 @@ export default {
   VerifyingContractFunding: "Verificando o financiamento do contrato",
   VerifyingHedgeFndingProposal: "Verificando a proposta de financiamento de hedge",
   VerifyingLongFndingProposal: "Verificando a longa proposta de financiamento",
-  Version: "Versão",
+  Version: "Weergawe",
   ViewContract: "Ver contrato",
   ViewIgnoredPoints: "Ver pontos ignorados",
   ViewIgnoredTokens: "Ver tokens ignorados",
@@ -800,7 +800,7 @@ export default {
   WaitingSmartBchTransaction: "Esperando pela transação em Smart BCH",
   Waiting_SEP20_PointSent: "Esperando o ponto de sep20 ser enviado",
   Waiting_SEP20_TokenSent: "Esperando o token de setembro ser enviado",
-  Wallet: "Carteira",
+  Wallet: "Beursie",
   WalletBalance: "Balanço da carteira",
   WalletBalanceCap: "Balanço da carteira",
   WalletConnect: "Carteira Connect",
@@ -808,7 +808,7 @@ export default {
   WalletHash: "HASH da carteira",
   WalletInfo: "Informações da carteira",
   WalletName: "Nome da carteira",
-  Wallets: "Carteiras",
+  Wallets: "Beursies",
   Warning: "Aviso",
   Week: "semana",
   Weeks: "semanas",
@@ -820,5 +820,5 @@ export default {
   YouSend: "Você envia",
   YouWillReceive: "Você receberá",
   YourMnemonic: "Sua frase de backup mnemônica é",
-  or: "ou"
+  or: "of"
 }
