+++ src/i18n/es-ar/index.js
@@ -10,7 +10,7 @@ export default {
   Add: "Agregar",
   AddAll: "Añadir todo",
   AddAnotherRecipient: "Agregar otro destinatario",
-  AddFungibleCashPoint: "Agregar CashPoint Fungible",
+  AddFungibleCashPoint: "Agregar CashPoint fungible",
   AddFungibleCashToken: "Agregar CashToken Fungible",
   AddNewDevice: "Agrega un nuevo dispositivo",
   Add_SEP20_Point: "Agregar punto de sep20",
@@ -51,7 +51,7 @@ export default {
   AssetSwap: "Intercambio de activos",
   AssetValueNote: "El valor del activo se basa en los precios en el momento de la transacción",
   AssetsSwappedSuccesfully: "Activos cambiados con éxito",
-  BCHBullDescription: "Cree contratos de todos los años al instante con un proveedor de liquidez",
+  BCHBullDescription: "Cree contratos de todos los días al instante con un proveedor de liquidez",
   Back: "Atrás",
   BackToEdit: "Volver a editar",
   BackendDown: "Nuestro servidor de backend no es posible.Esto podría deberse a que su conexión a Internet o que nuestro servidor esté temporalmente abajo",
@@ -227,7 +227,7 @@ export default {
   EnterOneByOne: "Ingrese la frase de semilla uno por uno",
   EnterPin: "Ingrese su PIN",
   Enter_SEP20_ContractAddress: "Ingrese la dirección del contrato de SEP20",
-  Enter_SLP_PointId: "Ingrese la identificación del punto SLP",
+  Enter_SLP_PointId: "Ingrese la ID de punto SLP",
   Enter_SLP_TokenId: "Ingrese SLP Token ID",
   ErrAcceptingCallRequest: "Error a aceptar la solicitud de llamada",
   Error: "Error",
