+++ src/i18n/ha/index.js
@@ -127,7 +127,7 @@ export default {
   Clear: "Share",
   ClearCallRequests: "Bayyana buƙatun kira",
   ClearCallRequestsPrompt: "Cire duk buƙatun kira.Ka tabbata?",
-  ClearFilter: "Share tace",
+  ClearFilter: "Share tacewa",
   ClickToCopyAddress: "Danna don kwafa adireshi",
   ClickToReveal: "Danna don bayyana",
   Close: "Rufa",
@@ -384,7 +384,7 @@ export default {
   Minutes: "ƙanƙane",
   MnemonicBackupPhrase: "Tattaunawa Mnemonic",
   MnemonicBackupPhraseDescription1: "Dubawa sau biyu idan wannan ya dace da jumlar ajiyar ku ta Mnemmonic",
-  MnemonicBackupPhraseDescription2: "Rubuta akan takarda ko kuma ɗaukar hoto kuma ku riƙe shi wani wuri mai lafiya",
+  MnemonicBackupPhraseDescription2: "Rubuta akan takarda ko kuma ɗaukar hoto kuma ku riƙe shi wani wuri mai aminci",
   Monthly: "Na wata",
   Months: "watanni",
   MultipleRecipientsUnsupported: "Masu karɓa da yawa ba su goyi bayan",
@@ -502,12 +502,12 @@ export default {
   Range: "Iyaka",
   Raw: "Ɗanye",
   Receive: "Sama",
-  Received: "Samu",
+  Received: "Karɓa",
   ReceiverWarningText1: "A karo na farko da karbar kudi don wannan kadara?Don Allah",
   ReceiverWarningText2: "Ajiyayyen Serve Sentranka",
   ReceiverWarningText3: "kafin a ci gaba.",
   ReceivingAddress: "Karɓar adireshi",
-  Recipient: "Magajiya",
+  Recipient: "Magaji",
   Recipients: "Masu karɓa",
   Recover: "Mai da",
   RecoverGift: "Murmurewa kyauta",
