+++ src/i18n/index.js
@@ -1,46 +1,141 @@
 import enUS from './en-us'
+import de from './de'
+import zhTW from './zh-tw'
+import zhCN from './zh-cn'
 import es from './es'
 import esAR from './es-ar'
-import zhCN from './zh-cn'
-import zhTW from './zh-tw'
-import de from './de'
-import ha from './ha'
 import pt from './pt'
 import ptBR from './pt-br'
+import ha from './ha'
+import af from './af'
+import ceb from './ceb'
+import fr from './fr'
+import tl from './tl'
+import nl from './nl'
+import id from './id'
+import it from './it'
+import ja from './ja'
+import ko from './ko'
 
 
-export const supportedLangs = {
-  'en-us': 'English',
-  'es': 'Spanish',
-  'es-ar': 'ArgentinianSpanish',
-  'zh-tw': 'ChineseTraditional',
-  'zh-cn': 'ChineseSimplified',
-  'de': 'German',
-  'ha': 'Hausa',
-  'pt': 'Portuguese',
-  'pt-br': 'BrazilianPortuguese',
+/**
+ * 
+ * NOTE: these are the languages supported by translate-google npm package
+ * 
+ * COMMENT the uncommented are the ones supported by paytaca-pos at the moment
+ * 
+ * TO ADD A LANGUAGE:
+ * 1. add language here and uncomment in the object below
+ * 2. add language code in Translator.translate()#supportedLangs variable
+ * 3. add a folder on src/i18n/ and add an index.js file in the folder
+ * 
+ * 
+ */
+const supportedLangs = {
+  af: 'Afrikaans',
+  // sq: 'Albanian',
+  // ar: 'Arabic',
+  // hy: 'Armenian',
+  // az: 'Azerbaijani',
+  // eu: 'Basque',
+  // be: 'Belarusian',
+  // bn: 'Bengali',
+  // bs: 'Bosnian',
+  // bg: 'Bulgarian',
+  // ca: 'Catalan',
+  ceb: 'Cebuano',
+  // ny: 'Chichewa',
+  'zh-cn': 'Chinese Simplified',
+  'zh-tw': 'Chinese Traditional',
+  // co: 'Corsican',
+  // hr: 'Croatian',
+  // cs: 'Czech',
+  // da: 'Danish',
+  nl: 'Dutch',
+  en: 'English',
+  // eo: 'Esperanto',
+  // et: 'Estonian',
+  tl: 'Filipino',
+  // fi: 'Finnish',
+  fr: 'French',
+  // fy: 'Frisian',
+  // gl: 'Galician',
+  // ka: 'Georgian',
+  de: 'German',
+  // el: 'Greek',
+  // gu: 'Gujarati',
+  // ht: 'Haitian Creole',
+  ha: 'Hausa',
+  // haw: 'Hawaiian',
+  // iw: 'Hebrew',
+  // hi: 'Hindi',
+  // hmn: 'Hmong',
+  // hu: 'Hungarian',
+  // is: 'Icelandic',
+  // ig: 'Igbo',
+  id: 'Indonesian',
+  // ga: 'Irish',
+  it: 'Italian',
+  ja: 'Japanese',
+  // jw: 'Javanese',
+  // kn: 'Kannada',
+  // kk: 'Kazakh',
+  // km: 'Khmer',
+  ko: 'Korean',
+  // ku: 'Kurdish (Kurmanji)',
+  // ky: 'Kyrgyz',
+  // lo: 'Lao',
+  // la: 'Latin',
+  // lv: 'Latvian',
+  // lt: 'Lithuanian',
+  // lb: 'Luxembourgish',
+  // mk: 'Macedonian',
+  // mg: 'Malagasy',
+  // ms: 'Malay',
+  // ml: 'Malayalam',
+  // mt: 'Maltese',
+  // mi: 'Maori',
+  // mr: 'Marathi',
+  // mn: 'Mongolian',
+  // my: 'Myanmar (Burmese)',
+  // ne: 'Nepali',
+  // no: 'Norwegian',
+  // ps: 'Pashto',
+  // fa: 'Persian',
+  // pl: 'Polish',
+  pt: 'Portuguese',
+  // ma: 'Punjabi',
+  // ro: 'Romanian',
+  // ru: 'Russian',
+  // sm: 'Samoan',
+  // gd: 'Scots Gaelic',
+  // sr: 'Serbian',
+  // st: 'Sesotho',
+  // sn: 'Shona',
+  // sd: 'Sindhi',
+  // si: 'Sinhala',
+  // sk: 'Slovak',
+  // sl: 'Slovenian',
+  // so: 'Somali',
+  es: 'Spanish',
+  // su: 'Sudanese',
+  // sw: 'Swahili',
+  // sv: 'Swedish',
+  // tg: 'Tajik',
+  // ta: 'Tamil',
+  // te: 'Telugu',
+  // th: 'Thai',
+  // tr: 'Turkish',
+  // uk: 'Ukrainian',
+  // ur: 'Urdu',
+  // uz: 'Uzbek',
+  // vi: 'Vietnamese',
+  // cy: 'Welsh',
+  // xh: 'Xhosa',
+  // yi: 'Yiddish',
+  // yo: 'Yoruba',
+  // zu: 'Zulu'
 }
-
-/*
-  NOTE:
-
-  Update the following files when adding a new language:
-
-  1. this file (see code above) 
-
-  2. components/settings/LanguageSelector.vue
-    = defaultLocaleOptions (variable)
-
-  3. pages/registration/accounts.vue
-    = supportedLangs (variable)
-  
-  4. i18n/translate.js
-    = supportedLangs (variable)
-
-  5. Add language name to group of texts on i18n/translate.js (words variable)
-    = words or phrases variable
-    e.g. Chinese: "Chinese"
-*/
 
 export default {
   'en-us': enUS,
@@ -48,8 +143,17 @@ export default {
   'zh-tw': zhTW,
   'es-ar': esAR,
   'pt-br': ptBR,
-  es,
   de,
-  ha,
+  es,
   pt,
+  ha,
+  af,
+  ceb,
+  nl,
+  id,
+  it,
+  ja,
+  ko,
+  fr,
+  tl,
 }
