+++ src/i18n/pt-br/index.js
@@ -29,7 +29,7 @@ export default {
   Amount: "Quantia",
   AmountGreaterThanBalance: "O valor é maior que o seu equilíbrio",
   AnyHedgeContract: "AnyendeDGECONTRATRA",
-  AnyHedgeNoExactMatchInfo: "Se não houver correspondência exata encontrada no pool, uma lista de ofertas semelhantes é sugerida",
+  AnyHedgeNoExactMatchInfo: "Se não houver uma correspondência exata encontrada no pool, uma lista de ofertas semelhantes é sugerida",
   AppInfo: "Informações do aplicativo",
   Applications: "Formulários",
   Approve: "Aprovar",
@@ -167,7 +167,7 @@ export default {
   CreateHedge: "Criando oferta de posição de hedge",
   CreateHedgeError: "Erro encontrado na criação de posição de hedge",
   CreateLong: "Criando oferta de posição longa",
-  CreateLongError: "Erro encontrado na criação de posição longa",
+  CreateLongError: "Erro encontrado na criação de uma posição longa",
   CreateNewCampaign: "Crie uma nova campanha",
   CreateNewWallet: "Crie uma nova carteira",
   CreateOrImportWallet: "Criar/importar carteira",
@@ -599,7 +599,7 @@ export default {
   SelectActiveSession: "Selecione sessão ativa",
   SelectAssetToBeReceived: "Selecione ativo a ser recebido",
   SelectAssetToSend: "Selecione ativo para enviar",
-  SelectBCHDenomination: "Denominação de BCH",
+  SelectBCHDenomination: "Denominação BCH",
   SelectCampaign: "Selecione a campanha",
   SelectCollection: "Selecione coleção",
   SelectCustomPoint: "Selecione o ponto personalizado",
