+++ src/i18n/zh-cn/index.js
@@ -232,7 +232,7 @@ export default {
   ErrAcceptingCallRequest: "错误接受呼叫请求",
   Error: "错误",
   EstimatedFee: "估计费用",
-  ExceededBalanceError: "超过余额{spendableBchValue}",
+  ExceededBalanceError: "超过余额​​{spendableBchValue}",
   ExpirationUpdated: "到期更新",
   Expired: "已到期",
   ExpiresAt: "到期",
@@ -737,7 +737,7 @@ export default {
   TransactionDeadline: "交易截止日期",
   TransactionHistory: "交易历史记录",
   TransactionId: "事务ID",
-  TransactionSent: "交易发送",
+  TransactionSent: "发送的交易",
   Transactions: "交易",
   Type: "类型",
   UTXOScanComplete: "UTXO扫描完成",
